python-ai/requirements.txt
paho-mqtt==2.1.0
pymongo==4.15.3
python-dotenv==1.2.1
pyttsx3==2.98
numpy>=1.24,<3
tensorflow>=2.20,<2.22
matplotlib>=3.8,<4
